requirements.txt
pytest
pytest-cov
codecov
pyyaml
mock